Can Compile code, changes everywhere

+++ headers/data.h
@@ -15,11 +15,11 @@ typedef struct mydata{
 
 
 
-int dist_msr(Data array,int index_a, int index_b);
+float dist_msr(Data array,int index_a, int index_b);
 
-int dist_manh(Data array,int index_a, int index_b);
+float dist_manh(Data array,int index_a, int index_b);
 
-int *get_weights(int *array, int data_of_interest, Data dataset, int array_lenth, int (* weight)(Data , int a, int b));
+float *get_weights(int *array, int data_of_interest, Data dataset, int array_lenth, float (* weight)(Data , int a, int b));
 
 void *import_data(char* filename, int *size);
 // void *import_data_bin(char * filename, int data_size){
